feat: support envd server destroy env
