+++ pkg/lang/ir/supervisor.go
@@ -36,7 +36,6 @@ working-directory = "${%[3]s}"
 
 [environment]
 keep-env = true
-re-export = [ "PATH", "SHELL", "USER", "%[3]s", "ENVD_AUTHORIZED_KEYS_PATH", "ENVD_HOST_KEY" ]
 
 [restart]
 strategy = "on-failure"
@@ -47,6 +46,15 @@ attempts = 5
 wait = "5s"
 `
 )
+
+func (g Graph) installHorust(root llb.State) llb.State {
+	horust := root.
+		File(llb.Copy(llb.Image(types.HorustImage), "/", "/usr/local/bin", llb.WithUIDGID(g.uid, g.gid)),
+			llb.WithCustomName("[internal] install horust")).
+		File(llb.Mkdir(types.HorustServiceDir, 0755, llb.WithParents(true), llb.WithUIDGID(g.uid, g.gid))).
+		File(llb.Mkdir(types.HorustLogDir, 0755, llb.WithParents(true), llb.WithUIDGID(g.uid, g.gid)))
+	return horust
+}
 
 func (g Graph) addNewProcess(root llb.State, name, command string) llb.State {
 	template := fmt.Sprintf(horustTemplate, name, command, types.EnvdWorkDir)
